Refactor AI bots: Replace bot_hard.js and normal_bot.js with Hard_bot.js and Medium_bot.js; update main.js to reflect new imports and functionality. Remove old bot files and ensure game logic is intact.
