+++ Hard_bot.js
